+++ package/lib/SimpleSchema_pick.tests.js
@@ -25,4 +25,15 @@ describe('SimpleSchema', function () {
     newSchema = schema.pick('blah');
     expect(Object.keys(newSchema.schema())).toEqual([]);
   });
+
+  it('error when you do not pick the parent', () => {
+    const schema = new SimpleSchema({
+      level1: { type: Object },
+      'level1.level2': { type: Boolean },
+    });
+
+    expect(() => {
+      schema.pick('level1.level2');
+    }).toThrow('"level1.level2" is in the schema but "level1" is not');
+  });
 });